+++ core/src/executive/internal_contract/abi/array.rs
@@ -1,3 +1,7 @@
+// Copyright 2020 Conflux Foundation. All rights reserved.
+// Conflux is free software and distributed under GNU General Public License.
+// See http://www.gnu.org/licenses/
+
 use super::{
     utils::{pull_slice, read_abi_list, ABIListWriter, LinkedBytes},
     ABIDecodeError, ABIVariable,
