+++ core/src/executive/internal_contract/abi/basic.rs
@@ -1,3 +1,7 @@
+// Copyright 2020 Conflux Foundation. All rights reserved.
+// Conflux is free software and distributed under GNU General Public License.
+// See http://www.gnu.org/licenses/
+
 use super::{utils::abi_require, ABIDecodeError, ABIVariable, LinkedBytes};
 use cfx_types::{Address, U256};
